Fix initial image height being undefined
This issue has been plaguing me for long enough to have almost quit the project. I was operating under an assumption that the nil pointer error I was getting MUST have been in reference to the options I was pointing in, but it was actually regarding the undefined image.

Resizing is still wonky for lack of a better term. Please fix and squash later :-)