WIP Stat displaying
Added very basic display of stats. meta.Data.chart refs statchart component which handles the display of stats. Currently displays upside-down, but need to figure out a whitespace solution to fix this.
Future reference it is possible to access a range of a slice by referencing the slice[lower:upper], which includes lower but excludes upper, so adding whitespace and then saving the tallest may be the play here.
Anyways please squash me later