Fix debug log outputting to stdout if no debug key
When I was looking for display bug I ended up going through the compiler warnings I fixed. When looking at the entrypoint I noticed the issue sprouted when I fixed the file. That fix included both ignoring File.Close errors explicitly in a deferred anonymous func and getting rid of the debug setenv. When I removed both fixes and saw the bug disappear I had just assumed it was some weird pointer logic in the closure instead of the default log output being pointed to std out instead of a file. This is why resizing fixed the UI. it refreshed the buffer.

There was an unfortunate vestigial logging statement in the stats display which triggered this bug and went uncaught by me when removing other vestigial features. Oopsy.

Anyways todo: create log package that checks debug before logging anyway, it just makes sense to init file when logging is attempted and debug is true since I know std out is not a viable target if it's false.